+++ zebra/connected.h
@@ -35,11 +35,11 @@ extern struct connected *connected_check_ptp(struct interface *ifp,
 					     union prefixconstptr d);
 
 extern void connected_add_ipv4(struct interface *ifp, int flags,
-			       struct in_addr *addr, uint8_t prefixlen,
+			       struct in_addr *addr, uint16_t prefixlen,
 			       struct in_addr *broad, const char *label);
 
 extern void connected_delete_ipv4(struct interface *ifp, int flags,
-				  struct in_addr *addr, uint8_t prefixlen,
+				  struct in_addr *addr, uint16_t prefixlen,
 				  struct in_addr *broad);
 
 extern void connected_delete_ipv4_unnumbered(struct connected *ifc);
@@ -49,10 +49,10 @@ extern void connected_down(struct interface *ifp, struct connected *ifc);
 
 extern void connected_add_ipv6(struct interface *ifp, int flags,
 			       struct in6_addr *address, struct in6_addr *broad,
-			       uint8_t prefixlen, const char *label);
+			       uint16_t prefixlen, const char *label);
 extern void connected_delete_ipv6(struct interface *ifp,
 				  struct in6_addr *address,
-				  struct in6_addr *broad, uint8_t prefixlen);
+				  struct in6_addr *broad, uint16_t prefixlen);
 
 extern int connected_is_unnumbered(struct interface *);
